--- internal/handlers/handlers.go
@@ -1,10 +1,0 @@
-// Package handlers provides handler functions for AKS MCP tools.
-package handlers
-
-// This file is kept for backward compatibility and package documentation.
-// Specific handler implementations have been moved to separate files based on resource type:
-// - cluster.go: Handlers for cluster-related operations
-// - vnet.go: Handlers for virtual network operations
-// - routetable.go: Handlers for route table operations
-// - nsg.go: Handlers for network security group operations
-// - common.go: Common utility functions used by multiple handlers
